llm-cps-ac/pyproject.toml
[tool.poetry]
name = "tb-optibot"
version = "0.1.0"
description = "TB OptiBot - Chatbot for control of heating system"
authors = ["Philippe Marziale <philippe.marziale@edu.hefr.ch>"]
license = "SPDX-License-Identifier: Apache-2.0"
readme = "README.md"
packages = [{include = "tb_optibot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
matplotlib = "^3.7.1"
pytz = "^2023.3"
openai = "^0.27.7"
langchain = "^0.0.215"
streamlit = "^1.23.1"
streamlit-chat = "^0.1.1"
faiss-cpu = "^1.7.4"
sqlalchemy = "^2.0.17"
tiktoken = "^0.4.0"
wikipedia = "^1.4.0"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
